refactored format code into separate module
