+++ arroyo-worker/Cargo.toml
@@ -15,6 +15,7 @@ arroyo-rpc = { path = "../arroyo-rpc" }
 arroyo-server-common = { path = "../arroyo-server-common" }
 arroyo-storage = { path = "../arroyo-storage" }
 arroyo-metrics =  { path = "../arroyo-metrics" }
+arroyo-formats = { path = "../arroyo-formats" }
 
 rand = "0.8"
 bincode = "2.0.0-rc.3"
