+++ arroyo-worker/src/connectors/fluvio/source.rs
@@ -1,6 +1,5 @@
 use crate::engine::{Context, StreamNode};
-use crate::formats::DataDeserializer;
-use crate::{SchemaData, SourceFinishType};
+use crate::{SourceFinishType};
 use anyhow::anyhow;
 use arroyo_macro::source_fn;
 use arroyo_rpc::formats::{Format, Framing};
@@ -20,6 +19,7 @@ use std::marker::PhantomData;
 use tokio::select;
 use tokio_stream::{Stream, StreamExt, StreamMap};
 use tracing::{debug, error, info, warn};
+use arroyo_formats::{DataDeserializer, SchemaData};
 
 use super::{FluvioTable, SourceOffset, TableType};
