+++ arroyo-worker/src/connectors/kafka/sink/test.rs
@@ -5,7 +5,7 @@ use std::io::Write;
 use std::time::{Duration, SystemTime};
 
 use crate::engine::{Context, OutQueue};
-use crate::SchemaData;
+use arroyo_formats::SchemaData;
 use arrow::datatypes::Field;
 use arroyo_rpc::formats::{Format, JsonFormat};
 use arroyo_types::CheckpointBarrier;
