+++ arroyo-worker/src/connectors/polling_http.rs
@@ -19,11 +19,11 @@ use arroyo_rpc::grpc::StopMode;
 use arroyo_state::tables::global_keyed_map::GlobalKeyedState;
 use tracing::{debug, info, warn};
 use typify::import_types;
+use arroyo_formats::{DataDeserializer, SchemaData};
 
-use crate::formats::DataDeserializer;
 use crate::{
     engine::{Context, StreamNode},
-    SchemaData, SourceFinishType,
+    SourceFinishType,
 };
 
 import_types!(schema = "../connector-schemas/polling_http/table.json");
