+++ build_dir/pipeline/Cargo.toml
@@ -18,6 +18,7 @@ arrow = { workspace = true}
 arrow-array = { workspace = true}
 arroyo-types = { path = "../../arroyo-types" }
 arroyo-worker = { path = "../../arroyo-worker" }
+arroyo-formats = { path = "../../arroyo-formats" }
 
 [package.metadata.wasm-pack.profile.release]
 wasm-opt = false
